refactor(dh): remove input drilling in actors audit log
