+++ libs/dh/market-participant/actors/feature-actors/src/lib/dh-actors-overview.component.html
@@ -47,8 +47,6 @@ limitations under the License.
 
     <dh-actors-table
       [tableDataSource]="tableDataSource"
-      [actorNumberNameLookup]="actorNumberNameLookup()"
-      [gridAreaCodeLookup]="gridAreaCodeLookup()"
       [isLoading]="isLoading()"
       [hasError]="hasError()"
     />
