+++ libs/dh/market-participant/actors/feature-actors/src/lib/drawer/actor-audit-log-tab/dh-actor-audit-log-tab.component.html
@@ -15,25 +15,25 @@ See the License for the specific language governing permissions and
 limitations under the License.
 -->
 <ng-container *transloco="let t; read: 'marketParticipant.actorsOverview.drawer'">
-  @if (auditLogLoading()) {
+  @if (isLoading()) {
     <vater-stack fill="horizontal" align="center">
       <watt-spinner />
     </vater-stack>
   } @else {
-    @if (auditLogDataSource.data.length > 0) {
+    @if (dataSource.data.length > 0) {
       <watt-card variant="solid">
         <watt-table
-          [dataSource]="auditLogDataSource"
-          [columns]="auditLogColumns"
+          [dataSource]="dataSource"
+          [columns]="columns"
           [hideColumnHeaders]="true"
           [suppressRowHoverHighlight]="true"
           sortBy="timestamp"
           sortDirection="desc"
         >
-          <ng-container *wattTableCell="auditLogColumns['timestamp']; let element">
+          <ng-container *wattTableCell="columns['timestamp']; let element">
             {{ element.timestamp | wattDate: "long" }}
           </ng-container>
-          <ng-container *wattTableCell="auditLogColumns['currentValue']; let entry">
+          <ng-container *wattTableCell="columns['currentValue']; let entry">
             @if (
               entry.change !== "CLIENT_SECRET_CREDENTIALS" &&
               entry.change !== "STATUS" &&
@@ -92,8 +92,8 @@ limitations under the License.
     } @else {
       <watt-empty-state
         size="small"
-        [icon]="auditLogHasError() ? 'custom-power' : 'cancel'"
-        [title]="auditLogHasError() ? t('tabs.shared.error') : t('tabs.shared.noData')"
+        [icon]="hasError() ? 'custom-power' : 'cancel'"
+        [title]="hasError() ? t('tabs.shared.error') : t('tabs.shared.noData')"
       />
     }
   }
