+++ libs/dh/market-participant/actors/feature-actors/src/lib/drawer/actor-audit-log-tab/dh-actor-audit-log-tab.component.ts
@@ -14,7 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Component, input, effect, computed, inject, ChangeDetectionStrategy } from '@angular/core';
+import {
+  Component,
+  input,
+  effect,
+  computed,
+  inject,
+  ChangeDetectionStrategy,
+  Signal,
+} from '@angular/core';
 import { TranslocoDirective, translate } from '@ngneat/transloco';
 
 import { VaterStackComponent } from '@energinet-datahub/watt/vater';
@@ -24,6 +32,11 @@ import { WattEmptyStateComponent } from '@energinet-datahub/watt/empty-state';
 import { WATT_CARD } from '@energinet-datahub/watt/card';
 import { WATT_TABLE, WattTableColumnDef, WattTableDataSource } from '@energinet-datahub/watt/table';
 import { dhActorAuditLogEntry } from '@energinet-datahub/dh/market-participant/actors/domain';
+import { query } from '@energinet-datahub/dh/shared/util-apollo';
+import {
+  GetActorsDocument,
+  GetGridAreasDocument,
+} from '@energinet-datahub/dh/shared/domain/graphql';
 
 import { DhActorAuditLogService } from '../dh-actor-audit-log.service';
 
@@ -46,31 +59,41 @@ import { DhActorAuditLogService } from '../dh-actor-audit-log.service';
 export class DhActorAuditLogTabComponent {
   private auditLogService = inject(DhActorAuditLogService);
 
-  auditLogLoading = this.auditLogService.auditLogQuery.loading;
-  auditLogHasError = computed(() => this.auditLogService.auditLogQuery.error() !== undefined);
+  private gridAreasQuery = query(GetGridAreasDocument);
+  private actorsQuery = query(GetActorsDocument);
+
+  private gridAreaCodeLookup: Signal<{ [key: string]: string }> = computed(() =>
+    this.getGridAreaCodes()
+  );
+
+  private actorNumberNameLookup: Signal<{ [key: string]: { number: string; name: string } }> =
+    computed(() => this.getActorNumberNames());
+
+  private gridAreasQueryHasError = computed(() => this.gridAreasQuery.error() !== undefined);
+  private auditLogHasError = computed(
+    () => this.auditLogService.auditLogQuery.error() !== undefined
+  );
+
+  isLoading = computed(
+    () => this.auditLogService.auditLogQuery.loading() || this.gridAreasQuery.loading()
+  );
+  hasError = computed(() => this.auditLogHasError() || this.gridAreasQueryHasError());
 
   actorId = input.required<string>();
-  actorNumberNameLookup = input.required<{
-    [key: string]: {
-      number: string;
-      name: string;
-    };
-  }>();
-  gridAreaCodeLookup = input.required<{
-    [key: string]: string;
-  }>();
 
-  auditLogDataSource = new WattTableDataSource<dhActorAuditLogEntry>([]);
-  auditLogColumns: WattTableColumnDef<dhActorAuditLogEntry> = {
+  dataSource = new WattTableDataSource<dhActorAuditLogEntry>([]);
+  columns: WattTableColumnDef<dhActorAuditLogEntry> = {
     timestamp: { accessor: 'timestamp' },
     currentValue: { accessor: 'currentValue' },
   };
 
   constructor() {
     effect(() => {
-      const logs = structuredClone(this.auditLogService.auditLogQuery.data()?.actorAuditLogs ?? []);
+      const logEntries = structuredClone(
+        this.auditLogService.auditLogQuery.data()?.actorAuditLogs ?? []
+      );
 
-      this.auditLogDataSource.data = logs.reverse();
+      this.dataSource.data = logEntries.reverse();
     });
 
     effect(() => {
@@ -87,7 +110,7 @@ export class DhActorAuditLogTabComponent {
 
     if (!values) return {};
 
-    const [payloadGln, payloadStartsAt, payloadGridArea, payloadProcessType, payloadStopsAt] =
+    const [payloadGln, payloadStartsAt, payloadGridAreaId, payloadProcessType, payloadStopsAt] =
       values;
     const actorNumberName = this.actorNumberNameLookup()[payloadGln];
 
@@ -95,11 +118,40 @@ export class DhActorAuditLogTabComponent {
       auditedBy: payload.auditedBy,
       actor: `${actorNumberName.number} • ${actorNumberName.name}`,
       startsAt: wattFormatDate(payloadStartsAt),
-      gridArea: this.gridAreaCodeLookup()[payloadGridArea],
+      gridArea: this.gridAreaCodeLookup()[payloadGridAreaId],
       processType: translate(
         'marketParticipant.actorsOverview.drawer.tabs.history.processTypes.' + payloadProcessType
       ),
       stopsAt: payloadStopsAt == undefined ? undefined : wattFormatDate(payloadStopsAt),
     };
+  }
+
+  private getActorNumberNames() {
+    const actors = this.actorsQuery.data()?.actors ?? [];
+
+    return actors.reduce(
+      (acc, actor) => {
+        acc[actor.id] = {
+          number: actor.glnOrEicNumber,
+          name: actor.name,
+        };
+
+        return acc;
+      },
+      {} as { [key: string]: { number: string; name: string } }
+    );
+  }
+
+  private getGridAreaCodes() {
+    const gridAreas = this.gridAreasQuery.data()?.gridAreas ?? [];
+
+    return gridAreas.reduce(
+      (acc, gridArea) => {
+        acc[gridArea.id] = gridArea.code;
+
+        return acc;
+      },
+      {} as { [key: string]: string }
+    );
   }
 }
