+++ libs/dh/market-participant/actors/feature-actors/src/lib/table/dh-actors-table.component.html
@@ -61,8 +61,4 @@ limitations under the License.
   }
 </vater-flex>
 
-<dh-actor-drawer
-  (closed)="onClose()"
-  [actorNumberNameLookup]="actorNumberNameLookup()"
-  [gridAreaCodeLookup]="gridAreaCodeLookup()"
-/>
+<dh-actor-drawer (closed)="onClose()" />
